Change Row::serialize_row signature so that type hinting is independent of struct members declaration order
When struct fields are not declared in the same order as in the database, type hints were passed to the wrong fields.
This affects only structs with fields requiring a type hint (decimal, bytes...) and results in an error from the server.

The `Row::serialize_row` method signature is

```rust
    fn serialize_row(
        self,
        type_hints: &[&Type]
    ) -> Result<Vec<(Cow<'static, str>, Value)>>;
```

It is called in `client.rs` as follows:

```rust
let types = first_block.column_types.values().collect::<Vec<_>>();
rows.into_iter()
    .map(|x| x.serialize_row(&types[..]))
```

The `Row` derive macro implements `serialize_row` by iterating over fields, and calling `types[idx]` to get the type hint for each field
There is however no guarantee that the fields order will match between the ones in the struct declaration and in `first_block` (returned by the server).

The solution is pretty simple but requires a change to the `Row` trait:
In `client.rs`, `column_types` is originally an `IndexMap` in a `Block`. Rather than passing it as a slice, one should pass the `IndexMap`, allowing to safely retrieve field type hints by name.

This will require a minor version bump.

The new test fails on master, but passes with this commit.