fix formatting of (U)Int{128,256} `Value`s
Large integers are parsed as floating point numbers by clickhouse by
default. If one wants to query e.g. `UInt128` columns, and the number
one wants to look up is large, this leads to no results being found,
because klickhouse currently outputs the number directly (causing
clickhouse to parse it as a float) if one uses `Value` to construct the
query.

Clickhouse has a feature where numbers can be suffixed with the type,
i.e., `123::UInt128`, to force it to not be converted to a float, but
stay a `UInt128`. See
https://github.com/ClickHouse/ClickHouse/issues/38480 as well.

This commit changes the formatting of `Value`, so that suffixes are
appended for `Int128` etc types.